refactor: convert poller delivery to TypeScript
